Add editor.cpp
